+++ src/Instance/index.js
@@ -6,3 +6,4 @@ export * from './InstanceDetails';
 export * from './InstanceMovement';
 export * from './InstanceMarc';
 export * from './InstanceEdit';
+export * from './ViewRequests';