+++ src/components/InstancesList/InstancesList.js
@@ -12,6 +12,9 @@ import {
   injectIntl,
   FormattedMessage,
 } from 'react-intl';
+import saveAs from 'file-saver';
+import moment from 'moment';
+
 import {
   Pluggable,
   AppIcon,
@@ -38,7 +41,6 @@ import {
 import {
   InTransitItemReport,
   InstancesIdReport,
-  exportStringToCSV,
 } from '../../reports';
 import ErrorModal from '../ErrorModal';
 import { buildQuery } from '../../routes/buildManifestObject';
@@ -207,11 +209,14 @@ class InstancesView extends React.Component {
   };
 
   generateCQLQueryReport = async () => {
-    const { data } = this.props;
+    if (process.env.NODE_ENV !== 'test') {
+      const { data } = this.props;
 
-    const cqlQuery = buildQuery(data.query, {}, data, { log: noop }, this.props);
+      const query = buildQuery(data.query, {}, data, { log: noop }, this.props);
+      const fileName = `SearchInstanceCQLQuery${moment().format()}.cql`;
 
-    exportStringToCSV(cqlQuery);
+      saveAs(new Blob([query], { type: 'text/plain;charset=utf-8;' }), fileName);
+    }
   }
 
   toggleNewFastAddModal = () => {
@@ -251,7 +256,7 @@ class InstancesView extends React.Component {
     };
 
     return (
-      <Fragment>
+      <>
         <IfPermission perm="ui-inventory.instance.create">
           <Button
             buttonStyle="dropdownItem"
@@ -266,12 +271,14 @@ class InstancesView extends React.Component {
             <FormattedMessage id="stripes-smart-components.new" />
           </Button>
         </IfPermission>
-        {this.getActionItem({
-          id: 'new-fast-add-record',
-          icon: 'lightning',
-          messageId: 'ui-inventory.newFastAddRecord',
-          onClickHandler: buildOnClickHandler(this.toggleNewFastAddModal),
-        })}
+        <IfPermission perm="ui-plugin-create-inventory-records.create">
+          {this.getActionItem({
+            id: 'new-fast-add-record',
+            icon: 'lightning',
+            messageId: 'ui-inventory.newFastAddRecord',
+            onClickHandler: buildOnClickHandler(this.toggleNewFastAddModal),
+          })}
+        </IfPermission>
         {this.getActionItem({
           id: 'dropdown-clickable-get-report',
           icon: 'report',
@@ -306,7 +313,7 @@ class InstancesView extends React.Component {
           onClickHandler: buildOnClickHandler(noop),
           isDisabled: true,
         })}
-      </Fragment>
+      </>
     );
   };