+++ src/routes/index.js
@@ -6,3 +6,4 @@ export { default as CreateItemRoute } from './CreateItemRoute';
 export { default as CreateHoldingRoute } from './CreateHoldingRoute';
 export { default as InstanceMarcRoute } from './InstanceMarcRoute';
 export { default as InstanceEditRoute } from './InstanceEditRoute';
+export { default as ViewRequestsRoute } from './ViewRequestsRoute';